LayoutTests/editing/undomanager/undomanager-reenter-expected.txt

Ensure that undoManager throws an exception when we re-enter it by calling undoManager functions during callbacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


call undoManager.transact() during the execute callback
PASS undoManager.transact(transaction) threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
call undoManager.undo() during the undo callback
PASS undoManager.undo() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
call undoManager.clearRedo() during the redo callback
PASS undoManager.redo() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS successfullyParsed is true

TEST COMPLETE
